+++ packages/gov-ui-kit/src/components/cards/cardSummary/index.ts
@@ -1,0 +1,2 @@
+export { CardSummary } from './cardSummary';
+export { type ICardSummaryProps } from './cardSummary.api';