+++ packages/gov-ui-kit/src/components/index.ts
@@ -2,6 +2,7 @@ export * from './actionItem';
 export * from './alerts';
 export * from './avatars';
 export * from './button';
+export * from './cards';
 export * from './icon';
 export * from './illustrations';
 export * from './input';