+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Implement `Card` and `CardSummary` components
+
 ### Changed
 
 -   Update `Spinner` and `Button` components to handle responsive sizes