+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.7] - 2024-01-12
+
 ### Added
 
 -   Implement `Card`, `CardSummary`, `Switch`, `Toggle` and `ToggleGroup` components